RF: Move anatomical apply into separate module
